api/web/requirements.txt
Django==1.8.1
gunicorn==19.3.0
psycopg2==2.6
djangorestframework
django-cors-headers